Accidentally wiped out my previous changes
Shouldn't have forked from the original master again. Anyway... Here are the changes again. And I added the little addition that I made to change the video_type to 'episode' which I did in service.py as well. Haven't checked why it worked without the addition yet as finishing an episode clearly didn't.